requirements.txt
asgiref==3.7.2
Django==5.0.1
django-filter==24.1
pillow==10.3.0
psycopg2==2.9.9
sqlparse==0.4.4
tzdata==2023.4
